requirements.txt
appnope==0.1.0
arcgis==1.6.2
attrs==19.1.0
autopep8==1.4.4
backcall==0.1.0
bleach==3.1.0
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.3
djangorestframework==3.10.1
djangorestframework-gis==0.14
entrypoints==0.3
gis==0.2.1
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
ipywidgets==7.5.0
jedi==0.14.1
Jinja2==2.10.1
json5==0.8.5
jsonschema==3.0.1
jupyter-client==5.3.1
jupyter-core==4.5.0
jupyterlab==1.0.2
jupyterlab-server==1.0.0
keyring==19.0.2
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
notebook==6.0.0
numpy==1.16.4
pandas==0.25.0
pandocfilters==1.4.2
parso==0.5.1
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycodestyle==2.5.0
Pygments==2.4.2
pyparsing==2.4.1.1
pyrsistent==0.15.3
pyshp==1.2.11
python-dateutil==2.8.0
pytz==2019.1
pyzmq==18.0.2
Send2Trash==1.5.0
six==1.12.0
sqlparse==0.3.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.0
